requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-allauth==65.10.0
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
